mcp-discord-server/pyproject.toml
[project]
name = "mcp-discord-server"
version = "0.1.0"
description = "A MCP server project"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
 "mcp>=1.6.0",
 "discord.py>=2.3.0",
 "audioop-lts; python_version >= '3.13'",
 "python-dotenv>=1.1.0",
]
[[project.authors]]
name = "Muhammad"
email = "hamzaadil56@gmail.com"

[build-system]
requires = [ "hatchling",]
build-backend = "hatchling.build"

[project.scripts]
mcp-discord-server = "mcp_discord_server:main"
